Consolidate private methods for managing .talismanrc
Authored-by: Owen Nelson <owen.nelson@thoughtworks.com>
